+++ src/Component/SearchBox.js
@@ -1,77 +1,57 @@
 import '../Assets/Style/SelectionMap.css';
 import React from 'react';
 import{ makeStyles } from "@material-ui/core/styles";
-import {Button,Grid,FormControl,FormControlLabel,Switch, RadioGroup, Radio, Slider, TextField,MenuItem, Paper, Typography} from "@material-ui/core";
+import {Button,Grid,FormControl,Switch,TextField, Paper, Typography} from "@material-ui/core";
+import SliderENABLE from "../Component/SliderENABLE";
+import SliderDISABLE from "../Component/SliderDISABLE";
+import TextFieldPays from "../Component/TextFieldPays";
+import TextFieldPaysDISABLED from "../Component/TextFieldPaysDISABLED";
+import RadioSearch from "../Component/RadioSearch";
+import RadioSearchDISABLED from "../Component/RadioSearchDISABLED";
+import DepartChoix from "../Component/DepartChoix";
+import DepartChoixDISABLE from "../Component/DepartChoixDISABLE";
+import DemandeChoix from "../Component/DemandeChoix";
+import DemandeChoixDISABLE from "../Component/DemandeChoixDISABLE";
 
-function valuetext(value) {
-    return `${value}`;
+
+function sliderState(bool){
+    if (bool === true){
+        return (<SliderENABLE/>);
+    } else {
+        return (<SliderDISABLE/>);
+    }
 }
-const currencies = [
-    {
-        value: 'Choisir',
-        label: 'Choisir un département',
-    },
-    {
-        value: 'TC',
-        label: 'Télécomunications, Services et Usages',
-    },
-    {
-        value: 'BIM',
-        label: 'Bioscience BIM',
-    },
-    {
-        value: 'BB',
-        label: 'Biosciences BB',
-    },
-    {
-        value: 'GCU',
-        label: 'Génie Civil et Urbanisme',
-    },
-    {
-        value: 'GEn',
-        label: 'Génie Energétique et Environnement',
-    },
-    {
-        value: 'GM',
-        label: 'Génie Mécanique',
-    },
-    {
-        value: 'IF',
-        label: 'Informatique',
-    },
-    {
-        value: 'GE',
-        label: 'Génie Electrique',
-    },
-    {
-        value: 'GI',
-        label: 'Génie Industriel',
-    },
-    {
-        value: 'SGM',
-        label: 'Science et Génie des Matériaux',
-    },
-];
 
-const currencies2 = [
-    {
-        value: 'Choisir',
-        label: 'Choisir un niveau de demande',
-    },
-    {
-        value: 'Faible',
-        label: 'Demande Faible',
-    },
-    {
-        value: 'Moyen',
-        label: 'Demande Moyenne',
-    },
-    {
-        value: 'Fort',
-        label: 'Demande Forte',
-    },
+function textFieldState(bool){
+    if (bool === true){
+        return (<TextFieldPays/>);
+    } else {
+        return (<TextFieldPaysDISABLED/>);
+    }
+}
 
-];
+function radioState(bool){
+    if (bool === true){
+        return (<RadioSearch/>);
+    } else {
+        return (<RadioSearchDISABLED/>);
+    }
+}
+function demandeState(bool){
+    if (bool === true){
+        return (<DemandeChoix/>);
+    } else {
+        return (<DemandeChoixDISABLE/>);
+    }
+}
+
+function departState(bool){
+    if (bool === true){
+        return (<DepartChoix/>);
+    } else {
+        return (<DepartChoixDISABLE/>);
+    }
+}
 
 const useStyles = makeStyles((theme) => ({
     paper: {
@@ -85,23 +65,6 @@ const useStyles = makeStyles((theme) => ({
 
 export default function SearchBox () {
     const classes = useStyles();
-
-    const [selectedValue, setSelectedValue] = React.useState('a');
-
-    const [currency, setCurrency] = React.useState('Choisir');
-
-    const handleChange = (event) => {
-        setCurrency(event.target.value);
-    };
-    const [currency2, setCurrency2] = React.useState('Choisir');
-
-    const handleChange3 = (event) => {
-        setCurrency2(event.target.value);
-    };
-
-    const handleChange2 = (event) => {
-        setSelectedValue(event.target.value);
-    };
 
     const [state, setState] = React.useState({
         checkedA: true,
@@ -159,16 +122,7 @@ export default function SearchBox () {
                             <Typography variant={'h6'}>Note globale</Typography>
                         </Grid>
                         <Grid item xs={6}>
-                            <Slider
-                                defaultValue={3}
-                                getAriaValueText={valuetext}
-                                aria-labelledby="discrete-slider"
-                                valueLabelDisplay="auto"
-                                step={1}
-                                marks
-                                min={0}
-                                max={5}
-                            />
+                            {sliderState(state.checkedA)}
                         </Grid>
                     </Grid>
 
@@ -187,16 +141,7 @@ export default function SearchBox () {
                             <Typography variant={'h6'}>Sécurité</Typography>
                         </Grid>
                         <Grid item xs={6}>
-                            <Slider
-                                defaultValue={3}
-                                getAriaValueText={valuetext}
-                                aria-labelledby="discrete-slider"
-                                valueLabelDisplay="auto"
-                                step={1}
-                                marks
-                                min={0}
-                                max={5}
-                            />
+                            {sliderState(state.checkedB)}
                         </Grid>
                     </Grid>
 
@@ -214,16 +159,7 @@ export default function SearchBox () {
                             <Typography variant={'h6'}>Coût de la vie</Typography>
                         </Grid>
                         <Grid item xs={6}>
-                            <Slider
-                                defaultValue={3}
-                                getAriaValueText={valuetext}
-                                aria-labelledby="discrete-slider"
-                                valueLabelDisplay="auto"
-                                step={1}
-                                marks
-                                min={0}
-                                max={5}
-                            />
+                            {sliderState(state.checkedC)}
                         </Grid>
                     </Grid>
 
@@ -241,16 +177,7 @@ export default function SearchBox () {
                             <Typography variant={'h6'}>Vie nocturne</Typography>
                         </Grid>
                         <Grid item xs={6}>
-                            <Slider
-                                defaultValue={3}
-                                getAriaValueText={valuetext}
-                                aria-labelledby="discrete-slider"
-                                valueLabelDisplay="auto"
-                                step={1}
-                                marks
-                                min={0}
-                                max={5}
-                            />
+                            {sliderState(state.checkedD)}
                         </Grid>
                     </Grid>
 
@@ -262,21 +189,13 @@ export default function SearchBox () {
                                 name="checkedE"
                                 color="primary"
                                 inputProps={{ 'aria-label': 'secondary checkbox' }}
-                            />                        </Grid>
+                            />
+                        </Grid>
                         <Grid item xs={3}>
                             <Typography variant={'h6'}>Vie culturelle</Typography>
                         </Grid>
                         <Grid item xs={6}>
-                            <Slider
-                                defaultValue={3}
-                                getAriaValueText={valuetext}
-                                aria-labelledby="discrete-slider"
-                                valueLabelDisplay="auto"
-                                step={1}
-                                marks
-                                min={0}
-                                max={5}
-                            />
+                            {sliderState(state.checkedE)}
                         </Grid>
                     </Grid>
 
@@ -293,13 +212,14 @@ export default function SearchBox () {
                                 name="checkedF"
                                 color="primary"
                                 inputProps={{ 'aria-label': 'secondary checkbox' }}
-                            />                        </Grid>
+                            />
+                        </Grid>
                         <Grid item xs={3}>
                             <Typography variant={'h6'}>Pays</Typography>
                         </Grid>
                         <Grid item xs={3}>
                             <form className={classes.root} noValidate autoComplete="off">
-                                <TextField id="outlined-basic" label="Pays" variant="filled" size="small"  />
+                                {textFieldState(state.checkedF)}
                             </form>
                         </Grid>
                     </Grid>
@@ -313,17 +233,15 @@ export default function SearchBox () {
                                 name="checkedG"
                                 color="primary"
                                 inputProps={{ 'aria-label': 'secondary checkbox' }}
-                            />                        </Grid>
+                            />
+                        </Grid>
                         <Grid item xs={3}>
                             <Typography variant={'h6'}>Hors Europe</Typography>
                         </Grid>
                         <Grid item xs={6}>
                             <div>
                                 <FormControl component="fieldset">
-                                    <RadioGroup row aria-label="position" name="position" defaultValue="end">
-                                        <FormControlLabel value="end" control={<Radio color="primary" checked={selectedValue === 'a'} onChange={handleChange2} value="a" name="radio-button-demo" inputProps={{ 'aria-label': 'A' }}/>} label="Oui" />
-                                        <FormControlLabel value="end" control={<Radio color="primary" checked={selectedValue === 'b'} onChange={handleChange2} value="b" name="radio-button-demo" inputProps={{ 'aria-label': 'B' }}/>} label="Non" />
-                                    </RadioGroup>
+                                    {radioState(state.checkedG)}
                                 </FormControl>
                             </div>
                         </Grid>
@@ -338,25 +256,14 @@ export default function SearchBox () {
                                 name="checkedH"
                                 color="primary"
                                 inputProps={{ 'aria-label': 'secondary checkbox' }}
-                            />                        </Grid>
+                            />
+                        </Grid>
                         <Grid item xs={3}>
                             <Typography variant={'h6'}>Demande</Typography>
                         </Grid>
                         <Grid item xs={6}>
-                            <TextField
-                                id="filled-select-currency"
-                                select
-                                label="Demande"
-                                value={currency2}
-                                onChange={handleChange3}
-                                variant="filled"
-                            >
-                                {currencies2.map((option) => (
-                                    <MenuItem key={option.value} value={option.value}>
-                                        {option.label}
-                                    </MenuItem>
-                                ))}
-                            </TextField>
+                            {demandeState(state.checkedH)}
+
                         </Grid>
 
                     </Grid>
@@ -369,25 +276,13 @@ export default function SearchBox () {
                                 name="checkedI"
                                 color="primary"
                                 inputProps={{ 'aria-label': 'secondary checkbox' }}
-                            />                        </Grid>
+                            />
+                        </Grid>
                         <Grid item xs={3}>
                             <Typography variant={'h6'}>Département</Typography>
                         </Grid>
                         <Grid item xs={6}>
-                            <TextField
-                                id="filled-select-currency"
-                                select
-                                label="Département"
-                                value={currency}
-                                onChange={handleChange}
-                                variant="filled"
-                            >
-                                {currencies.map((option) => (
-                                    <MenuItem key={option.value} value={option.value}>
-                                        {option.label}
-                                    </MenuItem>
-                                ))}
-                            </TextField>
+                            {departState(state.checkedI)}
                         </Grid>
 
                     </Grid>
