+++ src/Component/SliderENABLE.js
@@ -1,0 +1,20 @@
+import {Slider} from "@material-ui/core";
+import React from "react";
+
+function valuetext(value) {
+    return `${value}`;
+}
+export default function SliderENABLE () {
+    return(
+        <Slider
+            defaultValue={3}
+            getAriaValueText={valuetext}
+            aria-labelledby="discrete-slider"
+            valueLabelDisplay="auto"
+            step={1}
+            marks
+            min={0}
+            max={5}
+        />
+    );
+}
