+++ src/Component/TextFieldPays.js
@@ -1,0 +1,14 @@
+import {TextField} from "@material-ui/core";
+import React from "react";
+
+
+export default function TextFieldPaysDISABLED () {
+    return(
+        <TextField
+            id="outlined-basic"
+            label="Pays"
+            variant="filled"
+            size="small"
+        />
+    );
+}
